requirements.txt
attrs==19.2.0
Babel==2.7.0
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-analytical==2.5.0
django-import-export==1.2.0
django-multiurl==1.3.0
django-phone-field==1.8.0
django-phonenumber-field==3.0.1
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4.1
odfpy==1.4.0
openpyxl==3.0.0
phonenumbers==8.10.20
Pillow==6.2.0
pkg-resources==0.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pytz==2019.2
PyYAML==5.1.2
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
